Bug 1744606 - Cast return type of MaxChannels() to int32_t r=padenot
Fix warning: narrowing conversion from 'uint32_t' (aka 'unsigned int') to signed type 'int32_t' (aka 'int') is implementation-defined [cppcoreguidelines-narrowing-conversions]

Depends on D133000

Differential Revision: https://phabricator.services.mozilla.com/D133001

UltraBlame original commit: cd912504b41615a8fc5906daefef0a3332d058d8

+++ dom/media/webrtc/MediaEngineWebRTCAudio.cpp
@@ -101,7 +101,7 @@ nsresult MediaEngineWebRTCMicrophoneSource::EvaluateSettings(
   
   
   
-  int32_t maxChannels = mDeviceInfo->MaxChannels();
+  int32_t maxChannels = static_cast<int32_t>(mDeviceInfo->MaxChannels());
